Add CreateDatabaseDiscoveryRuleStatementUpdater
