+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/main/java/org/apache/shardingsphere/proxy/backend/text/distsql/rdl/impl/RDLBackendHandler.java
@@ -20,7 +20,7 @@ package org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl;
 import org.apache.shardingsphere.governance.core.registry.config.event.rule.RuleConfigurationsAlteredSQLNotificationEvent;
 import org.apache.shardingsphere.infra.config.RuleConfiguration;
 import org.apache.shardingsphere.infra.config.scope.SchemaRuleConfiguration;
-import org.apache.shardingsphere.infra.distsql.RDLStatementUpdater;
+import org.apache.shardingsphere.infra.distsql.RDLUpdater;
 import org.apache.shardingsphere.infra.eventbus.ShardingSphereEventBus;
 import org.apache.shardingsphere.infra.metadata.resource.ShardingSphereResource;
 import org.apache.shardingsphere.infra.spi.ShardingSphereServiceLoader;
@@ -48,7 +48,7 @@ import java.util.stream.Collectors;
 public abstract class RDLBackendHandler<T extends SQLStatement, R extends SchemaRuleConfiguration> extends SchemaRequiredBackendHandler<T> {
     
     static {
-        ShardingSphereServiceLoader.register(RDLStatementUpdater.class);
+        ShardingSphereServiceLoader.register(RDLUpdater.class);
     }
     
     public RDLBackendHandler(final T sqlStatement, final BackendConnection backendConnection) {
@@ -60,7 +60,7 @@ public abstract class RDLBackendHandler<T extends SQLStatement, R extends Schema
     protected final ResponseHeader execute(final String schemaName, final T sqlStatement) {
         R currentRuleConfig = findCurrentRuleConfiguration(schemaName).orElse(null);
         try {
-            RDLStatementUpdater updater = TypedSPIRegistry.getRegisteredService(RDLStatementUpdater.class, sqlStatement.getClass().getCanonicalName(), new Properties());
+            RDLUpdater updater = TypedSPIRegistry.getRegisteredService(RDLUpdater.class, sqlStatement.getClass().getCanonicalName(), new Properties());
             updater.checkSQLStatement(schemaName, sqlStatement, currentRuleConfig, ProxyContext.getInstance().getMetaData(schemaName).getResource());
             updater.updateCurrentRuleConfiguration(schemaName, sqlStatement, currentRuleConfig);
         } catch (final ServiceProviderNotFoundException ignored) {
